solution/requirements/prod.txt
django==4.2.10
djangorestframework==3.14.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
django-filter==23.5
